% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feed_dat}
\alias{feed_dat}
\title{Feed Data}
\format{A data frame with columns:
\describe{
\item{ep_num}{Episode number}
\item{movie}{Movie title}
\item{pub_date}{Day episode was released}
\item{pub_datetime}{Date and time episode was released}
\item{duration}{Length of episode in minutes}
\item{description}{Description of the episode}
\item{explicit}{Explicit episode marker}
\item{keywords}{Episode keywords}
\item{scraped_date}{Date when RSS feed was scraped}
\item{full_title}{Original title from RSS feed}
}}
\source{
http://theflophouse.libsyn.com/rss
}
\usage{
feed_dat
}
\description{
Episode information taken from the podcast RSS feed.
}
\keyword{datasets}
