% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recs_dat}
\alias{recs_dat}
\title{Recommendations}
\format{A data frame with columns:
\describe{
\item{ep_num}{Episode number}
\item{host}{Host who made the recommendation}
\item{imdb_id}{IMDB ID}
\item{scraped_date}{Date when wiki was scraped}
}}
\source{
http://theflophouserecommends.herokuapp.com/
}
\usage{
recs_dat
}
\description{
IMDB IDs for recommendations made by the hosts
}
\keyword{datasets}
