% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_dat}
\alias{wiki_dat}
\title{Wiki Data}
\format{A data frame with 5 columns:
\describe{
\item{ep_num}{Episode number}
\item{final_judgement}{Summary of final judgements on the movie}
\item{hosts}{The hosts of the episode}
\item{notes}{Episode notes}
\item{scraped_date}{Date when wiki was scraped}
}}
\source{
\url{http://flophousepodcast.wikia.com/wiki/Episode_List}
}
\usage{
wiki_dat
}
\description{
Episode list taken from the Flophouse Wiki
}
\keyword{datasets}
